+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ContextPolicyTest.java
@@ -90,7 +90,6 @@ import com.oracle.truffle.api.nodes.IndirectCallNode;
 import com.oracle.truffle.api.nodes.RootNode;
 import com.oracle.truffle.api.test.CompileImmediatelyCheck;
 import com.oracle.truffle.api.test.common.AbstractExecutableTestLanguage;
-import com.oracle.truffle.api.test.common.TestUtils;
 import com.oracle.truffle.tck.tests.TruffleTestAssumptions;
 
 public class ContextPolicyTest {
@@ -131,7 +130,6 @@ public class ContextPolicyTest {
 
     @Registration(contextPolicy = ContextPolicy.SHARED)
     static class AssertTestLanguage extends AbstractExecutableTestLanguage {
-        static final String ID = TestUtils.getDefaultLanguageId(AssertTestLanguage.class);
 
         @Override
         @TruffleBoundary